+++ src/components/GifGridItem.jsx
@@ -6,7 +6,6 @@
     <div className="card">
       <img src={url} alt={title}></img>
       <p>{title}</p>
-      GifGridItem
     </div>
   )
 }
