+++ src/components/index.js
@@ -1,0 +1,4 @@
+
+export * from './AddCategory';
+export * from './GifGrid';
+export * from './GifGridItem';
